+++ doc/source/whats_new.rst
@@ -75,6 +75,8 @@ Changelog
 
    - Add support for SNR estimation by `Eric Larson`_
 
+   - Add support for Savitsky-Golay filtering of Evoked and Epochs by `Eric Larson`_
+
 BUG
 ~~~
 
@@ -893,7 +895,7 @@ of commits):
 
 .. _Daniel Strohmeier: http://www.tu-ilmenau.de/bmti/fachgebiete/biomedizinische-technik/dipl-ing-daniel-strohmeier/
 
-.. _Eric Larson: http://faculty.washington.edu/larsoner/
+.. _Eric Larson: http://larsoner.com
 
 .. _Denis Engemann: https://github.com/dengemann